+++ README.md
@@ -100,23 +100,58 @@ type Repository struct {
 }
 ```
 
-Installation
-------------
+CLI Installation
+----------------
+
+```sh
+$ go get github.com/ChimeraCoder/gojson/gojson
+```
+
+Assuming `$GOPATH/bin` is in your `PATH`, you can now invoke `gojson` directly.
+
+
+API Installation
+----------------
 
 ```sh
 $ go get github.com/ChimeraCoder/gojson
 ```
 
-Assuming `$GOPATH/bin` is in your `PATH`, you can now invoke `gojson` directly.
+Development
+-----------
+
+```
+$ git clone https://github.com/ChimeraCoder/gojson.git
+$ cd gojson
+$ go test
+```
+
+**Building CLI**
+
+```
+$ go build -o _build/gojson ./gojson
+```
+
+**Installing CLI**
+
+```
+$ go install ./gojson
+```
+
+**Formatting**
+
+```
+$ gofmt -w -e -s -l *.go **/*.go
+```
 
 Related Work
 ------------
 
-github.com/str1ngs/jflect 
+github.com/str1ngs/jflect
 
-License 
+License
 ----------
 
-gojson is free software distributed under Version 3 of the GNU Public License. 
+gojson is free software distributed under Version 3 of the GNU Public License.
 
 As of the time of writing, this is the same license used for gcc (and therefore gccgo), so it is unlikely to restrict use in any way. Note that the GPL does not extend to any output generated by gojson; the GPL only applies to software which includes copies of gojson itself.
