+++ json-to-struct_test.go
@@ -1,4 +1,4 @@
-package main
+package json2struct
 
 import (
 	"io/ioutil"
@@ -11,32 +11,32 @@ import (
 // It does not (yet) test for correctness of the end result
 func TestSimpleJson(t *testing.T) {
 	i := strings.NewReader(`{"foo" : "bar"}`)
-	if _, err := generate(i, "TestStruct", "main"); err != nil {
-		t.Error("generate() error:", err)
+	if _, err := Generate(i, "TestStruct", "main"); err != nil {
+		t.Error("Generate() error:", err)
 	}
 }
 
 // TestNullableJson tests that a null JSON value is handled properly
 func TestNullableJson(t *testing.T) {
 	i := strings.NewReader(`{"foo" : "bar", "baz" : null}`)
-	if _, err := generate(i, "TestStruct", "main"); err != nil {
-		t.Error("generate() error:", err)
+	if _, err := Generate(i, "TestStruct", "main"); err != nil {
+		t.Error("Generate() error:", err)
 	}
 }
 
 // TestSimpleArray tests that an array without conflicting types is handled correctly
 func TestSimpleArray(t *testing.T) {
 	i := strings.NewReader(`{"foo" : [{"bar": 24}, {"bar" : 42}]}`)
-	if _, err := generate(i, "TestStruct", "main"); err != nil {
-		t.Error("generate() error:", err)
+	if _, err := Generate(i, "TestStruct", "main"); err != nil {
+		t.Error("Generate() error:", err)
 	}
 }
 
 // TestInvalidFieldChars tests that a document with invalid field chars is handled correctly
 func TestInvalidFieldChars(t *testing.T) {
 	i := strings.NewReader(`{"f.o-o" : 42}`)
-	if _, err := generate(i, "TestStruct", "main"); err != nil {
-		t.Error("generate() error:", err)
+	if _, err := Generate(i, "TestStruct", "main"); err != nil {
+		t.Error("Generate() error:", err)
 	}
 }
 
@@ -52,7 +52,7 @@ func TestExample(t *testing.T) {
 		t.Error("error reading expected_output_test.go", err)
 	}
 
-	actual, err := generate(i, "User", "main")
+	actual, err := Generate(i, "User", "json2struct")
 	if err != nil {
 		t.Error(err)
 	}
